Sesion_1/01_vacuometro_py_V01/requirements.txt
pyserial
binascii
time

